+++ tests/combinators.test.js
@@ -1,5 +1,27 @@
 const QueryQl = require('./../lib/queryql');
 
+test('apply filters with default operator', () => {
+    var queryQl = new QueryQl()
+    queryQl.applyFilter({ field: '_embedded.relation.nick', value: '@sensorario' });
+    queryQl.applyFilter({ field: '_embedded.foo.bar.name', value: 'Simone' });
+    expect(queryQl.keepDefaultCombinator().getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&filtering[_embedded.relation.nick]=@sensorario'
+        + '&filtering[_embedded.foo.bar.name]=Simone'
+    );
+});
+
+test('apply filters with or filtering operator', () => {
+    var queryQl = new QueryQl()
+    queryQl.applyFilter({ field: '_embedded.relation.nick', value: '@sensorario' });
+    queryQl.applyFilter({ field: '_embedded.foo.bar.name', value: 'Simone' });
+    expect(queryQl.keepOrFilteringCombinator().getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&filtering_or[_embedded.relation.nick]=@sensorario'
+        + '&filtering_or[_embedded.foo.bar.name]=Simone'
+    );
+});
+
 test('apply filters with or operator', () => {
     var queryQl = new QueryQl()
     queryQl.applyFilter({ field: '_embedded.relation.nick', value: '@sensorario' });